mask and zero tensor dtype

+++ onnx_inference.py
@@ -143,7 +143,7 @@ class Llama:
         position_ids = cp.arange(seqlen, dtype=cp.int64).reshape((1, seqlen))
         position_ids[0][0] = pastlen
 
-        attention_mask = cp.ones((1, seqlen + pastlen), dtype=cp.float32)
+        attention_mask = cp.ones((1, seqlen + pastlen), dtype=cp.float16)
         attention_mask = self._prepare_decoder_attention_mask(
             attention_mask, (1, seqlen), hidden, pastlen)
 
@@ -152,7 +152,7 @@ class Llama:
             past_value = self.pastvalues[idx]
 
             if past_key is None:
-                zero_tensor = cp.zeros((1, 32, 0, 128), dtype=cp.float32)
+                zero_tensor = cp.zeros((1, 32, 0, 128), dtype=cp.float16)
                 inputs = {
                     'hidden_in': hidden,
                     'attn_mask': attention_mask,
@@ -204,7 +204,7 @@ class Llama:
         position_ids = cp.arange(seqlen, dtype=cp.int64).reshape((1, seqlen))
         position_ids[0][0] = pastlen
 
-        attention_mask = cp.ones((1, seqlen + pastlen), dtype=cp.float32)
+        attention_mask = cp.ones((1, seqlen + pastlen), dtype=cp.float16)
         attention_mask = self._prepare_decoder_attention_mask(attention_mask, (1, seqlen), hidden, pastlen)
 
         for idx in range(self.DECODER_COUNT):
@@ -212,7 +212,7 @@ class Llama:
             past_value = self.pastvalues[idx]
 
             if past_key is None:
-                zero_tensor = cp.zeros((1, 32, 0, 128), dtype=cp.float32)
+                zero_tensor = cp.zeros((1, 32, 0, 128), dtype=cp.float16)
                 inputs = {
                     'hidden_in': hidden,
                     'attn_mask': attention_mask,
